Refactor rendering code: split large files, deduplicate lighting constants, remove unnecessary partial
Split NifGeometryExtractor.cs (2,275 lines) into 4 focused files:
- NifSceneGraphWalker.cs (scene graph traversal)
- NifSkinningExtractor.cs (skeletal skinning)
- NifBlockParsers.cs (NIF block parsing utilities)
- NifGeometryExtractor.cs (public API only, 277 lines)

Split NifSpriteRenderer.cs (1,246 lines) into 4 files:
- NifTriangleData.cs (data types)
- NifScanlineRasterizer.cs (scanline rasterizer)
- NifPostProcessing.cs (bloom post-processing)
- NifSpriteRenderer.cs (render orchestration, 574 lines)

Split NpcAppearanceResolver.cs (1,267 lines) into 3 files:
- NpcEsmRecordParsers.cs (ESM record processing)
- NpcAppearanceScanEntries.cs (scan entry data models)
- NpcAppearanceResolver.cs (resolve logic, 654 lines)

Split RenderNifProcessor.cs (708 lines) into 2 files:
- RenderNifHelpers.cs (utilities and data types)
- RenderNifProcessor.cs (processing logic, 489 lines)

Also includes: extract RenderLightingConstants.cs (shared between CPU/GPU renderers),
remove unnecessary partial from RenderNifProcessor, move SkipNiObjectNET to NifBlockParsers,
extract NifAnimationParser.cs, split RenderNpcCommand into NpcHeadBuilder/NpcBodyBuilder/
NpcRenderHelpers, and miscellaneous ESM/runtime/CLI improvements.

+++ src/FalloutXbox360Utils/Core/Formats/Esm/Conversion/Processing/EsmRecordParser.cs
@@ -164,7 +164,7 @@ internal static class EsmRecordParser
 
             // Validate: parse as GRUP header, GroupSize must be >= 24 and fit in file
             var candidate = EsmParser.ParseRecordHeader(data.AsSpan(scan), bigEndian);
-            if (candidate != null && candidate.DataSize >= 24 && scan + (long)candidate.DataSize <= data.Length)
+            if (candidate != null && candidate.DataSize >= 24 && scan + candidate.DataSize <= data.Length)
             {
                 offset = scan;
                 return true;
